refactor: 영화 아이템 컴포넌트 함수 분리 및 함수명 변경
