+++ docs/REQUIREMENTS_2.md
@@ -58,7 +58,7 @@
 - [x] 이미지
 - [x] 인덱스
 - [x] 모달
-- [ ] 영화 아이템
+- [x] 영화 아이템
 - [ ] 영화 아이템
 
 ### 프로그래밍 요구사항
